% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covmat.R
\name{Covmat}
\alias{Covmat}
\alias{Covmat,SepGP-method}
\title{Covmat: Calculate the Covariance Matrix for SepGP}
\usage{
Covmat(object, D1, D2)

\S4method{Covmat}{SepGP}(object, D1, D2)
}
\arguments{
\item{object}{An object of class SepGP}

\item{D1}{First DOE}

\item{D2}{Second DOE}
}
\value{
Covariance matrix
}
\description{
This function computes the covariance matrix between two DOEs based on a SepGP model.
}
\examples{
# Test function
fxt <- function(x, t) {sin(2 * pi * t * (x - t))}
# Design matrix D
D <- matrix(seq(0, 1, length = 10), ncol = 1)
# Define evaluation points
tt <- seq(0, 1, length = 10)
# Compute FD
FD <- apply(array(tt), 1, function(t){apply(D, 1, function(x){fxt(x,t)})})
# Create a SepGP object
gp <- SepGP(D, FD, s = 2, covtype = "gauss")
Covmat(gp, D, D)
}
