% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSepGP.R
\name{predictSepGP}
\alias{predictSepGP}
\alias{predictSepGP,SepGP-method}
\title{predictSepGP: Make Predictions Using a SepGP Model}
\usage{
predictSepGP(object, Dpred, covcompute = TRUE)

\S4method{predictSepGP}{SepGP}(object, Dpred, covcompute = TRUE)
}
\arguments{
\item{object}{An instance of class \code{SepGP}, representing the Gaussian Process model fitted to training data.}

\item{Dpred}{Matrix containing new design points for which predictions are to be made. Each row represents a new input point in the same dimension as \code{D} in \code{SepGP}.}

\item{covcompute}{A logical value indicating whether the posterior covariance should be computed. The default value is \code{TRUE}.}
}
\value{
A list with the following elements:
\item{Mean}{Posterior mean vector of the predictions for each new input in \code{Dpred}.}
\item{Covsp}{Spatial posterior covariance matrix of the predictions.}
\item{Covmat}{Posterior covariance matrix of the predictions, quantifying the uncertainty for each point in \code{Dpred}.}
}
\description{
Compute predictions for new spatial points using a Gaussian Process model with a separable covariance structure (\code{SepGP}).
It computes the posterior mean and covariance matrix for the specified design points based on the model’s prior and covariance functions.
}
\examples{
# Test function
fxt <- function(x, t) {sin(2 * pi * t * (x - t))}
# Design matrix D
D <- matrix(seq(0, 1, length = 10), ncol = 1)
# Define evaluation points
tt <- seq(0, 1, length = 10)
# Compute FD
FD <- apply(array(tt), 1, function(t){apply(D, 1, function(x){fxt(x,t)})})
# Create a SepGP object
gp <- SepGP(D, FD, s = 2, covtype = "exp", nugget=1e-12)
# Calculate predictions for new design points
Dpred <- matrix(seq(0.1, 0.9, length = 5), ncol = 1)
Yp <- predictSepGP(gp, Dpred)
print(Yp$Mean)
print(Yp$Covmat)
}
