% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SepGP.R
\docType{class}
\name{SepGP-class}
\alias{SepGP-class}
\title{SepGP: Gaussian Processes with Separable Covariance Functions}
\description{
The \code{SepGP} class implements Gaussian processes with separable covariance functions for regression and interpolation tasks.
It provides flexibility in specifying covariance structures, priors, and additional regularization through the nugget effect, allowing efficient hyperparameter optimization.
}
\section{Slots}{

\describe{
\item{\code{D}}{Design of experiments matrix containing the input points.}

\item{\code{FD}}{Matrix representing evaluations of the computer code at the points in \code{D}.}

\item{\code{s}}{Degree of the polynomial used for the prior mean function.}

\item{\code{B}}{Matrix of coefficients for the prior mean, based on \code{D}.}

\item{\code{covtype}}{Type of covariance function; options include \code{"gauss"}, \code{"exp"}, \code{"matern3_2"}, and \code{"matern5_2"}.}

\item{\code{phi}}{Vector of hyperparameters controlling the covariance structure.}

\item{\code{philow}}{Lower bounds for the hyperparameters in \code{phi}.}

\item{\code{phiup}}{Upper bounds for the hyperparameters in \code{phi}.}

\item{\code{Sigmat}}{Noise covariance matrix, representing observation noise in the data.}

\item{\code{nugget}}{A small positive value added to the diagonal of the covariance matrix to improve numerical stability, particularly in cases with near-zero variance or when data points are very close to each other. Defaults to \code{1e-8}.}
}}

