% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SepGP.R
\name{SepGP}
\alias{SepGP}
\title{Constructor for SepGP Class}
\usage{
SepGP(
  D,
  FD,
  s,
  B = NULL,
  covtype = "gauss",
  phi = NULL,
  philow = NULL,
  phiup = NULL,
  Sigmat = NULL,
  nugget = 1e-16
)
}
\arguments{
\item{D}{Matrix representing the design of experiments (input points).}

\item{FD}{Matrix representing the computer code evaluations on \code{D}.}

\item{s}{Degree of the polynomial for the prior mean function.}

\item{B}{Optional. Coefficients for the prior mean on \code{D}. If not specified, defaults to a matrix of ones.}

\item{covtype}{Type of covariance function to be used, with default value \code{"gauss"}. Acceptable values are \code{"gauss"}, \code{"exp"}, \code{"matern3_2"}, and \code{"matern5_2"}.}

\item{phi}{Optional. Hyperparameter vector for the covariance function. If not specified, defaults to a vector of ones.}

\item{philow}{Optional. Lower bounds for each hyperparameter in \code{phi}. If not specified, defaults to \code{0.1} for each element.}

\item{phiup}{Optional. Upper bounds for each hyperparameter in \code{phi}. If not specified, defaults to \code{10} for each element.}

\item{Sigmat}{Optional. Temporal covariance matrix. If not specified, defaults to an identity matrix of appropriate size.}

\item{nugget}{Optional. A small positive value added to the diagonal of the covariance matrix to improve numerical stability, defaulting to \code{1e-8}.}
}
\value{
An object of class \code{SepGP} representing a Gaussian process model with the specified settings.
}
\description{
Constructs an instance of the \code{SepGP} class, initializing a Gaussian process model with customizable covariance, prior mean structures, and optional nugget effect.
}
\examples{
# Test function
fxt <- function(x, t) {sin(2 * pi * t * (x - t))}
# Design matrix D
D <- matrix(seq(0, 1, length = 10), ncol = 1)
# Define evaluation points
tt <- seq(0, 1, length = 10)
# Compute FD
FD <- apply(array(tt), 1, function(t){apply(D, 1, function(x){fxt(x,t)})})
# Create a SepGP object
gp <- SepGP(D, FD, s = 2, covtype = "gauss")
}
